# service/src/main/java/org/openinfralabs/caerus/service/model/Udf.java
package org.openinfralabs.caerus.service.model;

import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import lombok.Data;

import java.io.Serializable;
import java.math.BigDecimal;
import java.util.List;

@Data
@ApiModel(description = "Details about UDF configuration.")
public class Udf implements Serializable {

    @ApiModelProperty(notes = "The unique ID of the UDF generated by the system when first created.")
    private String id = null;
    @ApiModelProperty(notes = "The name of the UDF.")
    private String name = null;
    @ApiModelProperty(notes = "The package of the UDF, 0 = function and 1 = container.")
    private int pkg = 0; //0 = function; 1 = docker
    @ApiModelProperty(notes = "The programming language of the UDF.")
    private String language = null;
    @ApiModelProperty(notes = "The interface version of the UDF.")
    private BigDecimal interfaceVersion = null;
    @ApiModelProperty(notes = "The language version of the UDF.")
    private BigDecimal languageVersion = null;
    @ApiModelProperty(notes = "The main class to invoke the UDF.")
    private String main = null;
    @ApiModelProperty(notes = "The app code file name of the UDF, obtained by the system.")
    private String fileName = null;
    // optional
    private List<String> invocationEvents;


    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public int getPkg() {
        return pkg;
    }

    public void setPkg(int pkg) {
        this.pkg = pkg;
    }

    public String getLanguage() {
        return language;
    }

    public void setLanguage(String language) {
        this.language = language;
    }

    public BigDecimal getInterfaceVersion() {
        return interfaceVersion;
    }

    public void setInterfaceVersion(BigDecimal interfaceVersion) {
        this.interfaceVersion = interfaceVersion;
    }

    public BigDecimal getLanguageVersion() {
        return languageVersion;
    }

    public void setLanguageVersion(BigDecimal languageVersion) {
        this.languageVersion = languageVersion;
    }

    public String getMain() {
        return main;
    }

    public void setMain(String main) {
        this.main = main;
    }

    public String getFileName() {
        return fileName;
    }

    public void setFileName(String fileName) {
        this.fileName = fileName;
    }
    public List<String> getInvocationEvents() {
        return invocationEvents;
    }

    public void setInvocationEvents(List<String> invocationEvents) {
        this.invocationEvents = invocationEvents;
    }

}